backend-minimal/requirements.txt
fastapi==0.112.*
uvicorn[standard]==0.30.*
openai>=1.40.0
python-multipart==0.0.9
python-dotenv==1.*
